backend/requirements/lock.txt
amqp==5.3.1
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
black==26.1.0
cbor2==5.8.0
celery==5.6.2
cffi==2.0.0
cfgv==3.5.0
channels==4.3.2
channels_redis==4.3.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.13.2
cron_descriptor==2.0.6
cryptography==46.0.3
daphne==4.2.1
decorator==5.2.1
distlib==0.4.0
Django==6.0.1
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.2
django-migration-linter==6.0.0
django-ninja==1.5.3
django-simple-history==3.11.0
django-stubs==5.2.9
django-stubs-ext==5.2.9
django-timezone-field==7.2.1
django_celery_results==2.6.0
ecdsa==0.19.1
executing==2.2.1
factory_boy==3.3.3
Faker==40.1.2
filelock==3.20.3
gunicorn==24.1.1
hyperlink==21.0.0
identify==2.6.16
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kombu==5.6.2
librt==0.7.8
matplotlib-inline==0.2.1
msgpack==1.1.2
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
parso==0.8.5
pathspec==1.0.3
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg-binary==3.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyOpenSSL==25.3.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
pytokens==0.4.0
PyYAML==6.0.3
redis==7.1.0
rsa==4.9.1
ruff==0.14.14
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
toml==0.10.2
traitlets==5.14.3
Twisted==25.5.0
txaio==25.12.2
types-PyYAML==6.0.12.20250915
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
vine==5.1.0
virtualenv==20.36.1
wcwidth==0.4.0
whitenoise==6.11.0
zope.interface==8.2
